MS_Visualizador/requirements.txt
black==24.4.2
faker==18.13.0
flask==2.2.5
flask-cors
flask-jsonpify
openpyxl~=3.1.2
pandas~=2.2.2
pika==1.3.2
python-dotenv~=1.0.1
pywin32==305
pyodbc~=5.1.0
requests==2.31.0
SQLAlchemy~=2.0.29
waitress==3.0.0